58-oppdaterte-sparekalkulator

+++ core/src/main/java/it1901/SavingsCalc.java
@@ -12,7 +12,7 @@ public class SavingsCalc {
      * @param period 
      * @return the total amount after the given period.
      */
-    public static double calculation(int monthlyAmount, int lumpAmount, int period, double interestRate) {
+    public static int calculation(int monthlyAmount, int lumpAmount, int period, double interestRate) {
         if (period == 0) 
             period = 1;
         
@@ -31,7 +31,7 @@ public class SavingsCalc {
 
         double sum = lumpAmountSum + monthlyAmountSum;
 
-        return Math.round(sum);
+        return (int) Math.round(sum);
     }
