+++ marxbankfx/src/main/java/it1901/SavingsCalcController.java
@@ -59,7 +59,7 @@ public class SavingsCalcController {
     /**
      * Calculates the total amount based on the values in text fields
      */
-    private double calculateTotalAmount() {
+    private int calculateTotalAmount() {
         int ma = Integer.parseInt(monthlyAmount.getText());
         int la = Integer.parseInt(lumpAmount.getText());
         int p = Integer.parseInt(period.getText());
@@ -67,7 +67,7 @@ public class SavingsCalcController {
         
         double calc = SavingsCalc.calculation(ma, la, p, ir);
 
-        return Math.round(calc);
+        return (int) Math.round(calc);
     }
